+++ src/limits/fixed.rs
@@ -3,9 +3,9 @@ use async_trait::async_trait;
 use super::{LimitAlgorithm, Sample};
 
 /// A simple, fixed concurrency limit.
-pub struct Fixed(usize);
+pub struct Fixed(u32);
 impl Fixed {
-    pub fn new(limit: usize) -> Self {
+    pub fn new(limit: u32) -> Self {
         assert!(limit > 0);
 
         Self(limit)
@@ -14,10 +14,10 @@ impl Fixed {
 
 #[async_trait]
 impl LimitAlgorithm for Fixed {
-    fn limit(&self) -> usize {
+    fn limit(&self) -> u32 {
         self.0
     }
-    async fn update(&self, _reading: Sample) -> usize {
+    async fn update(&self, _reading: Sample) -> u32 {
         self.0
     }
 }
